Implement `COPY --from` caching.
Layers are created and cached when building the source stage and added
to the image when the target stage is being built. Cache key used for the
`COPY --from` layer is calculated based on the source stage and is later
added to the target layer cache key to ensure that changes in the layer
are properly detected.